Update SicCode to support 'Mixed Eligiblity'.
Mixed eligibility is where certain SIC codes are 
not eligible when the business is in the public 
sector, but all other businesses are eligible.

The UI will warn the lender of this when they 
choose such an SIC code and will have to decide
whether to proceed with the loan application.