Add rake task to safely export data
We are dealing with IL2 confidential data, and wish to be able to view
a suitably cleansed form of that data for working with locally.

This should allow us to work with production-like volumes of data, and
also troubleshoot specific issues that arise in production, without
requiring more people to have access to production.

For now, we are using a fork of the data-anonymization gem; there was a
problem with Rails mass-assignment stopping data being copied. Ideally
we will provide a suitable test case and fix to the upstream project and
then run off a published gem. This was the fastest way to try out a
working solution in production.