Quick PoC for streaming generation of CSV
We are seeing timeouts at nginx since the CSV reports are taking too
long to generate for production data.

This is a quick experiment to see whether we can use HTTP chunked
encoding to return a response so that Unicorn will stream the response
back to nginx and we won't get timeouts as a result.