Add error logging
Premium schedule report is currently blowing up when a collector
attempts to download a report.

Without production data, we are unable to reproduce this problem, so
this is an initial attempt to capture more details about the problem
and we can then have a conversation with CfE about what we might be
doing wrong.