Avoid parameterised extend
We don't want to extend instances at runtime based on request-time
parameters. It feels overly clever and (although not a concern in this
particular instance) will not be optimised very well by the VM so
your performance will take an unnecessary hit.

See http://blog.headius.com/2012/10/so-you-want-to-optimize-ruby.html
which talks about how Object#extend can be challenging.