Raise Devise timeout
We've had reports of people having problems with the 30 minute default
timeout.

Discussed with @jystewart; can't see anything wrong with bumping it for
now.

Possibly the UX needs looking at if people can't fill out and submit a
form in 30 minutes, but not a priority to look at that just now.